src/requirements.txt
llama-cpp-python
PyMuPDF
